+++ HISTORY.md
@@ -1,5 +1,7 @@
 ## ✒ 历史版本的特性介绍 (Features in old versions)
 
 ### v0.0.1
+
 > 此版本发布于 2021-xx-xx
-* 敬请期待。。。
+
+* 使用 yaml 配置创建 logit 的 Logger